requirements.txt
pillow==12.2.0
qrcode==8.2
fastapi==0.115.12
python-multipart==0.0.20
mangum==0.19.0
boto3==1.38.12
aws-cdk-lib==2.189.1
constructs>=10.0.0,<11.0.0
